# CsvWriter performance optimization