--- a/src/CsvHelper/Expressions/ObjectRecordWriter.cs
+++ b/src/CsvHelper/Expressions/ObjectRecordWriter.cs
@@ -27,11 +27,9 @@ namespace CsvHelper.Expressions
 		/// that will write the given record using the current writer row.
 		/// </summary>
 		/// <typeparam name="T">The record type.</typeparam>
-		/// <param name="record">The record.</param>
-		protected override Action<T> CreateWriteDelegate<T>(T record)
+		/// <param name="type">The type for the record.</param>
+		protected override Action<T> CreateWriteDelegate<T>(Type type)
 		{
-			var type = Writer.GetTypeForRecord(record);
-
 			if (Writer.Context.Maps[type] == null)
 			{
 				Writer.Context.Maps.Add(Writer.Context.AutoMap(type));