--- a/src/CsvHelper/Expressions/PrimitiveRecordWriter.cs
+++ b/src/CsvHelper/Expressions/PrimitiveRecordWriter.cs
@@ -25,11 +25,9 @@ namespace CsvHelper.Expressions
 		/// that will write the given record using the current writer row.
 		/// </summary>
 		/// <typeparam name="T">The record type.</typeparam>
-		/// <param name="record">The record.</param>
-		protected override Action<T> CreateWriteDelegate<T>(T record)
+		/// <param name="type">The type for the record.</param>
+		protected override Action<T> CreateWriteDelegate<T>(Type type)
 		{
-			var type = Writer.GetTypeForRecord(record);
-
 			var recordParameter = Expression.Parameter(typeof(T), "record");
 
 			Expression fieldExpression = Expression.Convert(recordParameter, typeof(object));