@snaumenko-st 2024-04-19
Eliminating all remaining `string` allocations seems to me much more difficult. As almost half of the code needs to be rewritten from scratch. And of course we don't have convenient support for that in earlier versions of .net.
So this PR is mostly compromise.