@JoshClose 2024-04-23
Does this same issue happen when reading?