@snaumenko-st 2024-04-25
> Does this same issue happen when reading?

No, when reading it's much better, and It allocates only a few bytes per operation.