@snaumenko-st 2024-04-25
@JoshClose I've tried the same benchmarks, and the results are almost the same, except with `ExpandoObject`. However it makes no sense to us. 
Thanks a lot!

P.S. My `CsvReader` benchmarks before and after your latest fix show almost the same results. So, I'm not sure whether it improved anything.