@JoshClose 2024-04-25
I changed expando to use any `IDictionary<string, object>` now, so that should cover most dynamics. It'll fall back to `IDynamicMetaObjectProvider`, which is slower.

I added a fast dynamic object that is created instead of `ExpandoObject` when reading.