@JoshClose 2024-04-25
This will be in the next release 32.0.0 on NuGet.